+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -24,6 +24,7 @@
 #include "Rule16p4Check.h"
 #include "Rule16p5Check.h"
 #include "Rule16p7Check.h"
+#include "Rule17p12Check.h"
 #include "Rule17p1Check.h"
 #include "Rule17p8Check.h"
 #include "Rule18p5Check.h"
@@ -73,6 +74,8 @@ public:
         "misrac-rule-16p7");
     CheckFactories.registerCheck<Rule17p1Check>(
         "misrac-rule-17p1");
+    CheckFactories.registerCheck<Rule17p12Check>(
+        "misrac-rule-17p12");
     CheckFactories.registerCheck<Rule17p8Check>(
         "misrac-rule-17p8");
     CheckFactories.registerCheck<Rule18p5Check>(